Many code checks, working for up to two subjects?! I made many code checks today. The scr_model1 works great so far. The scr_model2 not so much. As soon as I add more than 2 subjects, everything goes amiss.
Peculiar things about the bug:
I remove the link between subjects (the hyperparameters) and let the model run for 3000 iterations (2500s) for 5 subjects. It should be identical to fitting 5 subjects independently.
Subject 1,3 and 5 are converging in this particular case, 2 and 4 not. The data look practically identical. I will use identical data in a second, just to stay sane (see also the plots in "plots")
But I guess the nonconvergence of the hyperparameter could be a result of some subjects converging, some not.

One important lesson I learned:
don't constrain parameters you know are true! I put the sigma to my simulation value of 0.06 and nothing converged, but I did not notice because of another bug. Just something to keep in mind, STAN wants to find the values ;)