Read data and fit single subject model
I read in the data of the baseline condition and fitted the model. See 2015-11-23_single_subject_datafit.html for an example subject.
The model is converging if you get it in the correct range, => Initial values are still very important. I think we do need to fit the
data first with a ML fit per subject and then let the model run.

I repaired the posterior predictive and it is now working.

There is some preprocessing going on e.g. subsampling, removing data after the last onset