played arround a little bit with bayesian ML optimization
The optimization converges also for the hierarchical model, but I did not have time yet to check whether it is a good solution.
The google group states that it will likely find a local maximum instead of the global, if the chains also have problems to converge.
But then, chains take ages and the optimizer lets it run in seconds.

I also updated the indexing things, current rstan allows for smarter indexing, but could not test it due to a rstan bug.

Forgot what the other changes are sorry